requirements.txt
Django==3.0.4
django-extensions==2.2.8
factory-boy==2.12.0
flake8==3.7.9
ipdb==0.13.2
isort==4.3.21
black==19.10b0
